+++ jubakit/classifier.py
@@ -168,7 +168,7 @@ class Config(GenericConfig):
 
   @classmethod
   def methods(cls):
-    return ['perceptron', 'PA', 'PA1', 'PA2', 'CW', 'AROW', 'NHERD', 'NN']
+    return ['perceptron', 'PA', 'PA1', 'PA2', 'CW', 'AROW', 'NHERD', 'NN', 'cosine', 'euclidean']
 
   @classmethod
   def _default_method(cls):
@@ -185,10 +185,18 @@ class Config(GenericConfig):
                     'AROW',
                     'NHERD', 'normal_herd'):
       return {'regularization_weight': 1.0}
+    elif method in ('cosine', 'euclidean'):
+      return {
+        'nearest_neighbor_num': 128,
+        'local_sensitivity': 1.0,
+      }
     elif method in ('NN', 'nearest_neighbor'):
       return {
         'method': 'euclid_lsh',
-        'parameter': {'hash_num': 64},
+        'parameter': {
+          'threads': -1,  # use number of logical CPU cores
+          'hash_num': 64,
+        },
         'nearest_neighbor_num': 128,
         'local_sensitivity': 1.0
       }